Time: 91 ms (30.40%) | Memory: 88.5 MB (31.83%) - LeetSync

+++ 1694-make-sum-divisible-by-p/make-sum-divisible-by-p.cpp
@@ -6,7 +6,7 @@ public:
         int n=nums.size();
         int ans=n;
         for(auto i:nums)suff+=i;
-        unordered_map<long long,int>mp;
+        unordered_map<int,int>mp;
         mp[0]=-1;
         for(int i=0;i<n;i++){
             prefsum+=nums[i];
